ngrok-python-agent-example/requirements.txt
# Web framework
flask==3.0.0

# ngrok tunnel
pyngrok==7.0.5

# Environment variables
python-dotenv==1.0.0

# OpenAI API (optional - for LLM integration)
openai==1.3.0

# Additional useful packages
requests==2.31.0
